Projection sample

+++ Src/Events/Inventory/CreditApplied.cs
@@ -2,10 +2,10 @@
 
 namespace Events.Inventory
 {
-    public class CreditApplied
+    public class CreditApplied 
     {
         public Guid GeneralLedgerEntryId { get; set; }
-        public decimal Amount { get; set; }
+        public int Amount { get; set; }
         public Guid SkuId { get; set; }
         public string Account { get; set; }
     }
