+++ Src/Events/Inventory/DebitApplied.cs
@@ -2,10 +2,10 @@
 
 namespace Events.Inventory
 {
-    public class DebitApplied
+    public class DebitApplied 
     {
         public Guid GeneralLedgerEntryId { get; set; }
-        public decimal Amount { get; set; }
+        public int Amount { get; set; }
         public string Account { get; set; }
         public Guid SkuId { get; set; }
     }
